[LAD] Code clean-up (1)
- Removed unused functions (getChildNode(), get_dist_config())
- Moved many helper functions to misc_helpers.py
- Other minor clean-ups
